refactor(settings): organize microphone controls
